# backend/src/main/java/com/thalassa/backend/controllers/EquipmentController.java
package com.thalassa.backend.controllers;

import com.thalassa.backend.dto.EquipmentRequest;
import com.thalassa.backend.dto.EquipmentResponse;
import com.thalassa.backend.services.EquipmentService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/equipment")
public class EquipmentController {

  private final EquipmentService equipmentService;

  public EquipmentController(EquipmentService equipmentService) {
    this.equipmentService = equipmentService;
  }

  /** PUT /api/equipment/{id} Actualiza nombre, potencia y horas de uso de un equipo. */
  @PutMapping("/{id}")
  public ResponseEntity<EquipmentResponse> updateEquipment(
      @PathVariable Long id, @Valid @RequestBody EquipmentRequest request) {
    return ResponseEntity.ok(equipmentService.updateEquipment(id, request));
  }

  /** DELETE /api/equipment/{id} Elimina un equipo del acuario. */
  @DeleteMapping("/{id}")
  public ResponseEntity<Void> deleteEquipment(@PathVariable Long id) {
    equipmentService.deleteEquipment(id);
    return ResponseEntity.noContent().build();
  }
}