# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Thalassa API
  description: |
    API REST del backend de Thalassa — gestión de acuarios marinos.
    Este archivo es la **fuente de verdad** del contrato API.
    Los DTOs del paquete `com.thalassa.backend.dto` se generan automáticamente
    a partir de este documento mediante `openapi-generator-maven-plugin`.
  version: 1.0.0
  contact:
    name: Thalassa Project
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/auth/register:
    post:
      tags: [Autenticación]
      summary: Registrar nuevo usuario
      description: Crea la cuenta con plan FREE. No emite JWT; el cliente debe llamar a /login a continuación.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Email o username ya registrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Autenticación]
      summary: Iniciar sesión
      description: Autentica con email + password y devuelve el access token JWT (15 min) y el refresh token opaco (30 días).
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Autenticación correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales incorrectas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Autenticación]
      summary: Renovar access token
      description: |
        Rota el refresh token y emite un nuevo access token JWT (15 min) y un nuevo refresh token (30 días).
        El refresh token recibido queda revocado inmediatamente.
        Si se detecta reutilización de un token ya revocado, se invalidan todos los tokens del usuario.
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens renovados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Refresh token inválido, expirado o ya utilizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Autenticación]
      summary: Cerrar sesión
      description: Revoca el refresh token. Idempotente — no falla si el token ya estaba revocado.
      operationId: logout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Sesión cerrada correctamente
  /api/auth/forgot-password:
    post:
      tags: [Autenticación]
      summary: Solicitar reseteo de contraseña
      description: |
        Genera un token de un solo uso (TTL 1 hora) y lo envía por email.
        Siempre devuelve 204, independientemente de si el email está registrado,
        para no revelar información sobre los usuarios existentes.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '204':
          description: Solicitud procesada (el link se enviará si el email existe)
  /api/auth/reset-password:
    post:
      tags: [Autenticación]
      summary: Resetear contraseña con token
      description: |
        Valida el token, actualiza la contraseña y revoca todos los refresh tokens
        del usuario (fuerza re-login en todos los dispositivos).
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Contraseña actualizada correctamente
        '400':
          description: Token inválido, expirado o ya utilizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Usuario ───────────────────────────────────────────────────────────────
...
  /api/equipment/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: [Equipamiento]
      summary: Actualizar equipo
      operationId: updateEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
      responses:
        '200':
          description: Equipo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Equipamiento]
      summary: Eliminar equipo
      operationId: deleteEquipment
      responses:
        '204':
          description: Equipo eliminado
        '404':
          description: Equipo no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Energía (sub-recurso de acuario) ─────────────────────────────────────
...
  /api/species:
    get:
      tags: [Catálogo de Especies]
      summary: Buscar especies
      description: |
        Busca por nombre común o científico (case-insensitive).
        Sin parámetro `search` devuelve el catálogo completo.
      operationId: searchSpecies
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de especies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpeciesCatalogResponse'
  /api/species/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Catálogo de Especies]
      summary: Detalle de especie
      operationId: getSpeciesById
      responses:
        '200':
          description: Detalle de la especie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeciesCatalogResponse'
        '404':
          description: Especie no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Export ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtenido en POST /api/auth/login
  schemas:
...
    SubscriptionPlan:
      type: string
      description: Plan de suscripción del usuario
      enum: [FREE, REEFMASTER]
...
    LivestockCategory:
      type: string
      description: Taxonomic category of the specimen
      enum: [FISH, CORAL, INVERTEBRATE]
    EquipmentCategory:
      type: string
      description: Functional category of the equipment
      enum: [LIGHT, PUMP, SKIMMER, HEATER, OTHER]
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          description: Email del usuario registrado
        password:
          type: string
          minLength: 1
          description: Contraseña del usuario
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT de acceso (TTL 15 min)
        refreshToken:
          type: string
          nullable: true
          description: Refresh token opaco (TTL 30 días). Solo presente en login y refresh.
        username:
          type: string
        email:
          type: string
        subscriptionPlan:
          $ref: '#/components/schemas/SubscriptionPlan'
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          description: Refresh token opaco recibido en el login
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          description: Email de la cuenta cuya contraseña se quiere resetear
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 1
          description: Token de reseteo recibido por email
        newPassword:
          type: string
          minLength: 8
          pattern: '^(?=.*[A-Za-z])(?=.*\d).{8,}'
          description: 'Nueva contraseña: mínimo 8 caracteres, al menos una letra y un número'
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
          description: Nombre de usuario único
        email:
          type: string
          format: email
          maxLength: 100
          description: Email único
        password:
          type: string
          minLength: 8
          pattern: '^(?=.*[A-Za-z])(?=.*\d).{8,}'
          description: 'Contraseña: mínimo 8 caracteres, al menos una letra y un número'

    # ── Usuario ───────────────────────────────────────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        subscriptionPlan:
          $ref: '#/components/schemas/SubscriptionPlan'
        electricityPriceKwh:
          type: number
          format: double
          nullable: true
        locale:
          type: string
          description: "Interface language: en | de | es"
          nullable: true
        temperatureUnit:
          type: string
          description: "C or F"
          nullable: true
        volumeUnit:
          type: string
          description: "L or GAL"
          nullable: true
...
    EquipmentRequest:
      type: object
      required: [name, powerWatts, hoursPerDay]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        powerWatts:
          type: integer
          format: int32
          minimum: 1
        hoursPerDay:
          type: number
          format: double
          minimum: 0.1
          maximum: 24.0
        category:
          $ref: '#/components/schemas/EquipmentCategory'
    EquipmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        powerWatts:
          type: integer
          format: int32
        hoursPerDay:
          type: number
          format: double
        category:
          $ref: '#/components/schemas/EquipmentCategory'
          nullable: true
...
    SpeciesCatalogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commonName:
          type: string
        scientificName:
          type: string
        category:
          $ref: '#/components/schemas/LivestockCategory'
        reefSafe:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── Scraper ───────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Descripción legible del error
        timestamp:
          type: string
          format: date-time
          description: Momento en que ocurrió el error (ISO 8601)